Week9/requirements.txt
autogen-agentchat
autogen-core
autogen-ext

openai
groq

python-dotenv

faiss-cpu

sentence-transformers
transformers
torch